+++ modules/forms.js
@@ -17,13 +17,18 @@ const form = {
                         const list = $$("datatable1");
                         const form = $$("form1");
                         const item_data = form.getValues();
-                        if (item_data.id && form.validate()) {
-                            list.updateItem(item_data.id, item_data);
-                            webix.message("Information was changed to datatable!");
-                        }
-                        else if (form.validate()){
-                            list.add(item_data);
-                            webix.message("Information was added to datatable!");
+                        if (form.validate()) {
+                            const item_data = form.getValues();
+                            if(item_data.id)
+                            {
+                                list.updateItem(item_data.id, item_data);
+                                webix.message("Information was changed to datatable!");
+                            }
+                            else
+                            {
+                                list.add(item_data);
+                                webix.message("Information was added to datatable!");
+                            }
                         }
                         else{
                             webix.message("Information wasn't added to datatable!");
