+++ task-webix.html
@@ -11,7 +11,7 @@
 </head>
 
 <body>
-    <script type="text/javascript" src="main.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 
 </html>
